Randevularim sayfasi html css kisimlari tamamlandi

+++ Models/Servis.cs
@@ -3,8 +3,8 @@
     public class Servis
     {
         public int Id { get; set; }
-        public string Name { get; set; } 
+        public string Name { get; set; }
         public int HizmetSuresi { get; set; }
-        public int HizmetFiyat { get; set; }
+        public decimal HizmetFiyat { get; set; } // Decimal türünde olmalı
     }
 }
